Replace more Cucumber features with feature specs
