+++ spec/features/count_spec.rb
@@ -1,16 +1,5 @@
 require 'spec_helper'
 
-def select_vacation(from, to=nil)
-  range = to ? (from..to) : (from..from)
-  range.each do |date|
-    page.find_by_id(date).click
-  end
-end
-
-def save_calendar
-  page.find_by_id('save').click
-end
-
 feature 'User sees the count being updated' do
 
   scenario 'on blank calendar', js: true do
